fix: tighten character catalog contract
